% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensky.R
\name{opensky_flights}
\alias{opensky_flights}
\title{Get flights within a time interval}
\usage{
opensky_flights(begin, end)
}
\arguments{
\item{begin}{Start of time interval as Unix timestamp (integer)}

\item{end}{End of time interval as Unix timestamp (integer).
Interval must be <= 2 hours.}
}
\value{
tibble: icao24, callsign, departure_airport, arrival_airport,
first_seen, last_seen
}
\description{
Get flights within a time interval
}
